src/lambda_functions/data_transformer/requirements.txt
pandas
numpy
boto3
psycopg2-binary
yfinance